practice-app/src/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
psycopg2-binary